+++ 2019/src/day5.rs
@@ -1,7 +1,7 @@
 use super::intcode::intcode::{call_intcode, State};
 
 #[aoc(day5, part1)]
-pub fn solve_part1(input: &str) -> i64 {
+pub fn solve_part1(input: &str) -> i128 {
     let mut state = State {
         instruction_pointer: 0,
         relative_pointer: 0,
@@ -15,8 +15,8 @@ pub fn solve_part1(input: &str) -> i64 {
 
     state.instructions = input
         .split(",")
-        .map(|input| input.parse::<i64>().unwrap())
-        .collect::<Vec<i64>>();
+        .map(|input| input.parse::<i128>().unwrap())
+        .collect::<Vec<i128>>();
 
     call_intcode(&mut state);
 
@@ -27,7 +27,7 @@ pub fn solve_part1(input: &str) -> i64 {
 }
 
 #[aoc(day5, part2)]
-pub fn solve_part2(input: &str) -> i64 {
+pub fn solve_part2(input: &str) -> i128 {
     let mut state = State {
         instruction_pointer: 0,
         relative_pointer: 0,
@@ -39,8 +39,8 @@ pub fn solve_part2(input: &str) -> i64 {
 
     state.instructions = input
         .split(",")
-        .map(|input| input.parse::<i64>().unwrap())
-        .collect::<Vec<i64>>();
+        .map(|input| input.parse::<i128>().unwrap())
+        .collect::<Vec<i128>>();
 
     state.input.push(5);
